feature: Добавить подборки

+++ main-service/src/main/java/ru/practicum/ewm/compilation/dto/NewCompilationDto.java
@@ -15,7 +15,7 @@ import java.util.List;
 @Setter
 @ToString
 public class NewCompilationDto {
-    private List<Integer> events;
+    private List<Long> events;
     private Boolean pinned;
     @NotNull(groups = Create.class)
     private String title;
